1). Replace the previous validators and add ErrorText class to the validators.dart 2). Refactor the EmailSignUpModel class to use the new validators and ErrorText class. 3). Refactor the EmailSignInModel class to use the new validators and ErrorText class. 4). Pass the validators to the EmailSignUpFormChangeNotifierBased class. 5). Reposition the AlreadyHaveAnAccountCheck Widget in the EmailSignUpFormChangeNotifierBased class. 6). Create and add the EmailVerification page to the EmailSignUpFormChangeNotifierBased class. 7). Add the ForgotPassword page to the EmailSignInFormChangeNotifierBased class. 8). Pass the validators to the EmailSignInFormChangeNotifierBased class. 9). Wrap the input fields in the EmailSignInFormChangeNotifierBased class with a form so as to validate. 10). Wrap the comment widget in the EditEntryPage class with a Form so as to validate. 11). Add a truncate method to the EntriesListTile class to truncate overflow labels. 12). Add a truncate method to the EntriesListItem class to truncate overflow labels. 13). Add a truncate method to the JobListTile class to truncate overflow labels. 14). Add letter spacing to CustomTextStyle class. 15). Add errorStyle to CustomTextField class. 16). Add errorStyle to CustomIconTextField class. 17) Rename CancelAndNextButon to CancelAndSignInButton. 18). Reduce the fontSize of the FormSubmitbutton class. 19). Add EditJobPageManager class. 20). Implement EditJobPageManager class in the EditJobPage class. 21). Refactor the AlreadyHaveAnAccoountCheck class. 22). Add isoCode property to CustomUserModel class. 23). Add truncate method and  refactor minor things in the AccountPage. 24) Refactor the PhonePage class to use Locale. 25). Pass the IsoCode through the OTPPage class, so as to store in Firestore Database. 26). Set isoCode Property to empty String when a user signs up with any method in the AuthBase class. 27). Initialize CountryCode in main.dart

+++ lib/app/home/job_entries/edit_entry_page.dart
@@ -4,6 +4,7 @@ import 'package:my_time_tracker/app/home/job_entries/date_time_picker.dart';
 import 'package:my_time_tracker/app/home/job_entries/format.dart';
 import 'package:my_time_tracker/app/home/models/entry.dart';
 import 'package:my_time_tracker/app/home/models/job.dart';
+import 'package:my_time_tracker/app/sign_in/components/validators.dart';
 import 'package:my_time_tracker/common_widgets/custom_text_field.dart';
 import 'package:my_time_tracker/common_widgets/custom_text_style.dart';
 import 'package:my_time_tracker/common_widgets/form_submit_button.dart';
@@ -49,6 +50,7 @@ class _EditEntryPageState extends State<EditEntryPage> {
   String _comment;
   bool isLoading = false;
   FocusNode _commentFieldNode, _submitButtonNode;
+  final _formKey = GlobalKey<FormState>();
 
   String get scaffoldContent {
     if (widget.entry != null) {
@@ -97,26 +99,28 @@ class _EditEntryPageState extends State<EditEntryPage> {
   }
 
   Future<void> _setEntryAndDismiss(BuildContext context) async {
-    try {
-      setState(() {
-        isLoading = true;
-      });
-      final entry = _entryFromState();
-      await widget.database.setEntry(entry);
-      print('PASSED22222!!!!');
-      Navigator.of(context).pop();
-      MyCustomSnackBar(
-          enabled: widget.entry == null ? true : false,
-          text: scaffoldContent,
-          onPressed: () => widget.database.deleteEntry(entry)).show(context);
-    } catch (e) {
-      setState(() {
-        isLoading = false;
-      });
-      PlatformExceptionAlertDialog(
-        title: 'Operation failed',
-        exception: e,
-      ).show(context);
+    if (_formKey.currentState.validate()) {
+      try {
+        setState(() {
+          isLoading = true;
+        });
+        final entry = _entryFromState();
+        await widget.database.setEntry(entry);
+        print('PASSED22222!!!!');
+        Navigator.of(context).pop();
+        MyCustomSnackBar(
+            enabled: widget.entry == null ? true : false,
+            text: scaffoldContent,
+            onPressed: () => widget.database.deleteEntry(entry)).show(context);
+      } catch (e) {
+        setState(() {
+          isLoading = false;
+        });
+        PlatformExceptionAlertDialog(
+          title: 'Operation failed',
+          exception: e,
+        ).show(context);
+      }
     }
   }
 
@@ -225,16 +229,21 @@ class _EditEntryPageState extends State<EditEntryPage> {
   }
 
   Widget _buildComment() {
-    return CustomTextField(
-      focusNode: _commentFieldNode,
-      labelText: 'Add a comment',
-      enabled: isLoading == false,
-      textInputAction: TextInputAction.done,
-      keyboardType: TextInputType.text,
-      maxLength: 40,
-      controller: TextEditingController(text: _comment),
-      onChanged: (comment) => _comment = comment,
-      onEditingComplete: () => _commentEditingComplete(),
+    return Form(
+      key: _formKey,
+      autovalidateMode: AutovalidateMode.onUserInteraction,
+      child: CustomTextField(
+        focusNode: _commentFieldNode,
+        labelText: 'Add a comment',
+        enabled: isLoading == false,
+        textInputAction: TextInputAction.done,
+        keyboardType: TextInputType.text,
+        maxLength: 40,
+        validator: CommentValidator(),
+        controller: TextEditingController(text: _comment),
+        onChanged: (comment) => _comment = comment,
+        onEditingComplete: () => _commentEditingComplete(),
+      ),
     );
   }
 }
